Import of TSTANLEY/Logger-Simple-1.09 from CPAN.
gitpan-cpan-distribution: Logger-Simple
gitpan-cpan-version:      1.09
gitpan-cpan-path:         TSTANLEY/Logger-Simple-1.09.tar.gz
gitpan-cpan-author:       TSTANLEY
gitpan-cpan-maturity:     released

+++ t/01load.t
@@ -9,5 +9,11 @@ ok($log->isa('Logger::Simple'),'Correct object created');
 my $logname=$log->get_log;
 ok($logname eq $logfile,'Log file is loaded');
 
-unlink $logfile;
+undef $log;
+
+if($^O eq 'MSWin32'){
+  system "C:\\Windows\\System32\\cmd.exe \/c del t\\logfile";
+}else{ 
+  unlink $logfile;
+}
